introduce config URL at build time and a "root" key in config file
The idea is that everything ical-related (i.e. everything that's expected to change often over the course of a year) is in one location, which can be updated indepentently of the actual code.

For this to work, I moved the "config file generation script" (generate-calendars-json.js)
to the repository that holds the ical files. every ical-related is now over there, except for one file : configUrl.json (the path to the config file)

when deploying, configUrl must be adjusted to wherever the calendars.json is generated
its default value is ok during development if the calendars.json file is pulled at the root of the project.

I also introduced a "root" key in config file, to factor out the common prefix the file names are now relative to this root key, which is itself relative to the config file